[SMH-21] refactor(design-system/text): Text 컴포넌트의 폴더 이름을 소문자로 수정
