+++ packages/design-system/src/components/index.tsx
@@ -2,7 +2,7 @@ import Input from './input';
 
 export default { Input };
 export { Select } from './select';
-export { Text } from './Text';
+export { Text } from './text';
 export { Tab } from './tab';
 export { Checkbox } from './checkbox';
 export { CheckboxGroup } from './checkbox-group';
