+++ packages/design-system/src/components/text/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Text } from './index';
+import { Text } from '.';
 import type { Meta, StoryObj } from '@storybook/react';
 import styled from '@emotion/styled';
